requirements.txt
certifi==2024.7.4
chardet==4.0.0
cmake==3.30.0
diffusers==0.12.1
eagerpy==0.30.0
efficientnet-pytorch==0.7.1
filelock==3.15.4
fsspec==2024.6.1
huggingface-hub==0.23.4
idna==2.10
importlib_metadata==8.0.0
Jinja2==3.1.4
joblib==1.4.2
lit==18.1.8
loggers==0.1.4
lossers==0.0.9
MarkupSafe==2.1.5
mpmath==1.3.0
natsort==8.4.0
networkx==3.1
numpy==1.24.4
opencv-python==4.10.0.84
packaging==24.1
pillow==10.4.0
protobuf==5.27.2
PyYAML==6.0.1
regex==2024.5.15
requests==2.25.1
safetensors==0.4.3
scikit-learn==1.3.2
scipy==1.9.1
sympy==1.13.0
tensorboardX==2.6.2.2
threadpoolctl==3.5.0
timm==0.6.12
tokenizers==0.19.1
torch==2.0.0+cu117
torchattacks==3.5.1
torchvision==0.15.1+cu117
tqdm==4.66.1
transformers==4.42.4
triton==2.0.0
typing_extensions==4.12.2
urllib3==1.26.19
zipp==3.19.2
